Add scripts to perform LNT submissions
Summary:
The scripts are an alternative to using `lnt runtest test-suite` to
submit benchmarks to an LNT server. Having simple scripts in the
test-suite allows submissions without having lnt installed and gives
added flexibility in cases where `lnt runtest` is too strict.

The output is compatible with the LNT 'nts' testsuite format. (Long term
we really should create a new lnt schema that better matches the output
instead of having translation tables mapping lit names to LNT nts
names).

A typical test-suite run may look like this:

    $ mkdir build
    $ cd build
    $ cmake -GNinja path/to/test-suite
    $ ninja
    $ lit -j1 -o lit.json .
    $ path/to/test-suite/utils/lntsubmit.py \
        -M name=x86_default_config \
        --run-info compilerinfo.json \
        --url http://localhost:8000/db_default/submitRun \
        lit.json

If this accepted the next step is to change `lnt runtest test-suite` to use the scripts coming with the test-suite to avoid code duplication.

Differential Revision: https://reviews.llvm.org/D34132